Create CCE_MC-1000.txt
Port for MC-1000 (1985), from Brazilian manufacturer CCE.
(Info in Portuguese: http://mc-1000.wikispaces.com)
The sprite system was changed a bit. Sprites are not stored as single strings. MC-1000's BASIC doesn't have INSTR(), so splitting a sprite into its rows would require a very slow logic with FOR...NEXT and MID$(). Instead the rows of all sprites are stored, each for itself, in a string array.
MC-1000 text screen is only 32 cols x 16 rows, so the castle was redrawn in a smaller shape.